+++ src/System.CommandLine/Builder/CommandLineBuilder.cs
@@ -27,7 +27,7 @@ namespace System.CommandLine.Builder
 
         internal Func<BindingContext, IHelpBuilder> HelpBuilderFactory { get; set; }
 
-        internal Option HelpOption { get; set; }
+        public Option HelpOption { get; internal set; }
 
         internal ValidationMessages ValidationMessages { get; set; }
